% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankSampling.R
\name{calculateRankSampling}
\alias{calculateRankSampling}
\title{Rank Sampling}
\usage{
calculateRankSampling(weights, featuresNames, weightPerGroup, cutOff = -1)
}
\arguments{
\item{weights}{of features list for each metric function for each dataset sample}

\item{featuresNames}{to be ranked}

\item{weightPerGroup}{of eachrank}

\item{cutOff}{: 0 (all features), k (k most ranked features), -1 (automatic k calculation)}
}
\value{
rank subset of features ordered by its relevance
}
\description{
Calculates Feature Ranking List sorted by its relevance
}
