% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightsSampling.R
\name{calculateWeightsSampling}
\alias{calculateWeightsSampling}
\title{Sample features weights}
\usage{
calculateWeightsSampling(dataset, RandomRows, functionToUse, features)
}
\arguments{
\item{dataset}{original}

\item{RandomRows}{N subset index rows}

\item{functionToUse}{statistical function to use as relevancy metric of each feature}

\item{features}{list of features names}
}
\value{
N x M list of weights
}
\description{
for each sample calculate feature weights based on functionTouse metrics
}
