% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.by.contrib.R
\name{cutoff.by.contrib}
\alias{cutoff.by.contrib}
\title{Cutoff by Contrib}
\usage{
cutoff.by.contrib(attrs, contrib = 1)
}
\arguments{
\item{attrs}{a data.frame containing features weights to be sorted and selected}

\item{contrib}{selection factor close to 1, subsets all features with contribution > uniform contribution * contrib}
}
\value{
featurenames subset, inverse ordered by its weights
}
\description{
cutoff.by.contrib is an alternative to Fselector cutoff.biggest.diff.
It selects all features that have contribution higher that an uniform contribution.
}
