% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRowSamples.R
\name{getNRandomsRowsSubsets}
\alias{getNRandomsRowsSubsets}
\title{Get N random Samples}
\usage{
getNRandomsRowsSubsets(nRows, size, iterations)
}
\arguments{
\item{nRows}{dataset number of rows}

\item{size}{dataset partition number of rows}

\item{iterations}{number of partitions (N)}
}
\value{
N dataset row indexes partitions
}
\description{
Get N random row indexes subsets
}
