% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForestImportance2.R
\name{RForestImportance2}
\alias{RForestImportance2}
\title{RForestImportance2}
\usage{
RForestImportance2(formula, dataset)
}
\arguments{
\item{formula}{dependent and independent variables (class ~ .)}

\item{dataset}{as specified in FSelector::random.forest.importance}
}
\value{
featurenames subset, ordered by its weights as specified in FSelector::random.forest.importance
}
\description{
FSelector::random.forest.importance default importance.type=1 mean decrease in node accuracy. This variante option is importance.type=2 mean decrease in node impurity
}
